requirements.txt
Django==1.9
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
gunicorn==19.9.0
markdown2==2.3.1
Pillow==5.3.0
psycopg2==2.7.5
PyJWT==1.4.0
whitenoise==4.1
dj-static==0.0.6
dj-database-url==0.5.0
six==1.11.0
